src/requirements.txt
rasterio
numpy
pyyaml